remove compiler warnings from Windows WAF build

+++ modules/c/nrt/source/Utils.c
@@ -467,7 +467,7 @@ static void normalize_dms(int* pDegrees, int* pMinutes, int* pSeconds)
 
     const int minutes_to_seconds = 60;
     const int degrees_to_seconds = 60 * minutes_to_seconds;
-    double total_seconds = (*pDegrees * degrees_to_seconds) + (*pMinutes * minutes_to_seconds) + *pSeconds;
+    int total_seconds = (*pDegrees * degrees_to_seconds) + (*pMinutes * minutes_to_seconds) + *pSeconds;
 
     *pDegrees = total_seconds / degrees_to_seconds;
     total_seconds -= (*pDegrees * degrees_to_seconds);
@@ -502,7 +502,7 @@ static void adjust_dms(int* pDegrees, int* pMinutes, int* pSeconds, char* pDir,
 NRTPROT(void) nrt_Utils_geographicLatToCharArray(int degrees, int minutes,
                                                  double seconds_, char *buffer7)
 {
-    int seconds = round(seconds_);
+    int seconds = (int) round(seconds_);
 
     normalize_dms(&degrees, &minutes, &seconds);
     const int max_degrees = 90;
@@ -540,7 +540,7 @@ NRTPROT(void) nrt_Utils_geographicLatToCharArray(int degrees, int minutes,
 NRTPROT(void) nrt_Utils_geographicLonToCharArray(int degrees, int minutes,
                                                  double seconds_, char *buffer8)
 {
-    int seconds = round(seconds_);
+    int seconds = (int)round(seconds_);
 
     normalize_dms(&degrees, &minutes, &seconds);
     const int max_degrees = 180;
@@ -675,4 +675,3 @@ NRTAPI(char*) nrt_strdup(const char* src)
     }
     return NULL;
 }
-
